+++ packages/patrol_cli/lib/src/crossplatform/app_options.dart
@@ -409,6 +409,7 @@ class WebAppOptions {
     this.globalTimeout,
     this.shard,
     this.headless,
+    this.browserArgs,
   });
 
   final FlutterAppOptions flutter;
@@ -429,6 +430,7 @@ class WebAppOptions {
   final int? globalTimeout;
   final String? shard;
   final String? headless;
+  final String? browserArgs;
 
   /// Translates these options into a proper flutter build invocation.
   List<String> toFlutterBuildInvocation() {