+++ packages/patrol_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fix running devtools after moving test to `patrol_test/`. (#2922) 
+
 ## 4.1.0
 
 - Fix bug when running web tests with relative imports from integration_test directory. (#2906)