+++ packages/patrol_cli/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix running devtools after moving test to `patrol_test/`. (#2922) 
 - Fix listen to stdout/stderr streams in buildApkConfigOnly to prevent hang on Windows. (#2905)
+- Fix bug that --exclude was not working for directories. (#2918)
 
 ## 4.1.0