+++ packages/patrol_cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix running devtools after moving test to `patrol_test/`. (#2922) 
 - Fix listen to stdout/stderr streams in buildApkConfigOnly to prevent hang on Windows. (#2905)
 - Fix bug that --exclude was not working for directories. (#2918)
+- Allow to pass additional web browser args through `web-browser-args`. (#2932) 
 
 ## 4.1.0